Initial checkin of Rhino JSR-223 adapter code.
This version is identical to the version in the scripting repository
as of current writing.

This code will not initially work, as I've relocated the toplevel.js
into com.sun.phobos.script.javascript.resources, and the code must be
adapted to match.

Also, you will need to add a user library called Rhino in order for
Eclipse to be able to build this successfully.