Apply some much-needed fixes to RhinoScriptEngine.
The fixes are described at:

  + https://scripting.dev.java.net/issues/show_bug.cgi?id=18
  + https://scripting.dev.java.net/issues/show_bug.cgi?id=19

Unfortunately, I'm not sure if scripting.dev.java.net will survive the
transition to the new Kenai-based java.net, so I may have to write more
details about these changes later on.

Also, fix the location of toplevel.js, as described in the last commit.