Hook the Rhino script engine into OQLEngine.
Since this project is designed to run on Java 6, we don't need to use
reflection to decide if JSR 223 is supported---it always is! But we do
need to check if the Rhino script engine is successfully detected.

Rather than allow a fallback to the JDK-bundled Rhino, this version
specifically requests "rhino-nonjdk". This makes it much more likely
that we'll get a version of Rhino that supports compiled mode.