Use ImmutableSet.Builder to make referrer set.
The previous approach essentially duplicated the way an ArrayList works,
and is worth replacing with ArrayList for that reason alone. However,
the referrer set is really a set, and worth changing to a Set (no usage
of getReferers(), including from hat.js, relies on it being a List).

In particular, ImmutableSet has the interesting property of preserving
insertion order. It also has a builder that uses ArrayList behind the
scenes. So, I've opted to use ImmutableSet and its builder, rather than
using ArrayList (or arrays!) manually.