Fix the reverse-order sorter functions.
Negating a number does not work correctly in two's-complement systems if
the number is the most-negative value. For example, -Integer.MIN_VALUE
has the same value as Integer.MIN_VALUE.

I don't believe this is a real issue in this case, but just to do the
Right Thing, I'm now using bitwise complement instead.