Add reference chain support for InstancesQuery.
Extract out the functionality that's now shared between HistogramQuery
and InstanceQuery into QueryHandler and Misc. Implement a "path"
property for queries that can be used to create self-links.

Split out the referee/referrer parameters in HistogramQuery---in
actuality, only the first class specified is a referee; the rest are
referrer classes. As a bonus, this simplifies the code a bit, because
we no longer need to special-case the first item in the reference
chain.