Implement more reference chaining.
Enable the referees of a reference chain to be inspected, also. (This
activates the links in the Reference Count column in the histogram.)

Also implement referrer chaining for refsByType. Referee chaining is
currently not implemented. Additionally, fix the way that reference
counting is done---the numbers should match up with what's seen in
the histogram.

Make the second parameter to ClassResolver say whether the class must
be specified, as opposed to whether the class is optional; I found the
latter more confusing.