Add preliminary support for language-specific models.
Currently, these support OpenJDK 6 and JRuby 1.2, since those are the
versions we currently use at On-Site. Also, I've only implemented the
model classes needed so far; I'll add more later on an as-needed basis.

Note that all fasthat code from On-Site are licensed under GPLv2 _OR
LATER_ (unlike OpenJDK), still with the Classpath Exception. This keeps
it compatible with the rest of the hat code, while enabling GPLv3 (or
later) projects to use the language-specific models independently of the
rest of the hat code.