Don't "lazily initialize" trivially-initialised Snapshot fields.
Generally, there's only one Snapshot object. Given that the newObjects
and siteTraces fields are initialised to empty, there's no savings (both
time-wise and memory-wise) in deferring their initialisation.

Also, make MyVisitor immutable. Modifying fields just to reuse the same
object is a bit of an antipattern these days, given that we have modern
GCs that are optimised for working with short-lived objects.