Implement a real language-specific class model.
This is the first substantial step towards implementing support for
language-specific histograms.

The changes in this commit are pretty significant: not only is there now
a class model (and the object model has been changed to work in terms of
class models instead of just class names), there's also a new JRubyClass
model that actually picks out the class name and properties the proper
way (i.e., what JRuby really does internally, not just some guess I was
working with), including walking through to the real class to get the
instance variable table.

This should generate much more useful output when listing instances
of RubyObject.

Internally, all the models now link back to the factory that created
them. This allows snapshot-global stuff, like class objects, to be used
from models.