Some peer review feedback and some other changes
Use Preconditions with the built in formatting rather than string concats.

I ran into a case where the input stream was null for a mustache template. I
suspect it was because I was rebuilding without restarting the server, so maybe
Java tried to load the resource after the jar had been re-written. Regardless,
the added precondition would make a repeat of that more easy to diagnose.